% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_raster.R
\name{align_raster}
\alias{align_raster}
\title{Aligns a raster to reference grid}
\usage{
align_raster(raster, temp_grid, method, clamp = FALSE, save_dir = NULL)
}
\arguments{
\item{raster}{SpatRaster, raster to process}

\item{temp_grid}{SpatRaster, template grid used for aligning (from createRastTemp() function)}

\item{method}{Character, method of interpolation. E.g. "bilinear", "near". See terra::project()}

\item{clamp}{Boolean or numeric vector, to clamp min and max after reprojecting. TRUE will automatically calcuate min max, or custom vector with min max values e.g. c(0,100)}

\item{save_dir}{Character, path to save the raster}
}
\value{
SpatRast, with the provided specifications taken from the reference grid
}
\description{
Reprojects, resamples, aligns, crops, and masks SpatRaster to template raster
}
\examples{
\dontrun{
library(terra)
library(sf)

# Create AOI
coords <- base::matrix(c(
     -405086.7, 181125.8,
     -405086.7, 184647.8,
     -400843.1, 184647.8,
     -400843.1, 181125.8,
     -405086.7, 181125.8), 
     ncol = 2, byrow = TRUE)
aoi <- sf::st_polygon(list(coords)) |>
     sf::st_sfc(crs = 32198) |>
     sf::st_sf(crs = 32198)

# Grab Quebec LULC link
COG_link <- query_lookup(subject="utilisation-territoire",year=2020)$file_path

# Load LULC within the AOI
raster <- load_raster(COG_link=COG_link,aoi=aoi)

# Create template for resampling
temp <- create_rast_template(aoi=aoi,crs=2961,res=50)

# Align raster to grid
aligned_raster <- align_raster(raster=raster,temp_grid=temp,method="near")

# Test alignment
terra::compareGeom(aligned_raster, temp, stopOnError=FALSE)
terra::compareGeom(raster, temp, stopOnError=FALSE)
}
}
