% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_code.R
\name{area_code}
\alias{area_code}
\title{Extract exact area}
\source{
https://gis.stackexchange.com/questions/433375/calculate-area-for-raster-in-r
}
\usage{
area_code(rast, unit = "m")
}
\arguments{
\item{rast}{Raster input with categories to calculate the area}

\item{unit}{Units passed to cellSize and calculate the area}
}
\value{
Data frame with the code name (first column) and the associated area (second column)
}
\description{
Extract exact area
}
\examples{
\dontrun{
f <- system.file("ex/elev.tif", package="terra")
r <- rast(f)
area_code(r)
}
}
