% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azcopy.R
\name{azcp}
\alias{azcp}
\title{azcopy wrapper function to ease the use and transfer files to or from network drives}
\usage{
azcp(
  net_drive_path,
  local_path,
  nd2local = FALSE,
  account,
  container,
  SASkey,
  drive = NULL
)
}
\arguments{
\item{net_drive_path}{network drive path (can point to a folder or a file)}

\item{local_path}{Path to local disk}

\item{nd2local}{Should the transfer be network disk --> local disk (TRUE) or local disk --> network disk (FALSE)}

\item{account}{Account information to connect to the drive (P drive is hnatprojets.file.core, S drive is hnatprojetsprive.file.core.)}

\item{container}{Container in which the data is stored (P drive is projet-et-database, S drive is fs-hnatprojetsprive)}

\item{SASkey}{SAS key to connect. Starts with sv... or ?sv (note that the '?' is added automatically)}

\item{drive}{Set information automatically for 'P' or 'S' drive, if KEYAZCPP (for P drive) and KEYAZCPS (for S drive) are set in .Renviron file (usethis::edit_r_environ()), automatically generates account, container and SASkey for P and S drive}
}
\value{
A command that can be pasted in shell to run.
}
\description{
azcopy wrapper function to ease the use and transfer files to or from network drives
}
\examples{
\dontrun{
azcp(net_drive_path = "P:/Projets/Actif/importat_project_folder/3-Analyses",
local_path = "C:/Projects/importat_project_folder/*",
nd2local = FALSE,
account = 'hnatprojets.file.core',
container = 'projet-et-database',
SASkey = 'secret-key-here')
}
}
