% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raster.R
\name{calculate_bbox}
\alias{calculate_bbox}
\title{Internal function calculate_bbox}
\usage{
calculate_bbox(aoi, to_wgs84 = FALSE)
}
\arguments{
\item{aoi}{sf geometry or SpatRaster, study region}

\item{to_wgs84}{Boolean, to transform bbox to EPSG 4326 or keep original CRS}
}
\value{
numerical vector, containing bounding box coordinates
}
\description{
Calculates the bounding box for an area of interest
}
