% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sas.R
\name{check_account_name}
\alias{check_account_name}
\title{Internal function check_account_name}
\usage{
check_account_name()
}
\value{
Error message if account name is not set
}
\description{
Function to check if the Habitat blob account name is present in .Renviron file
}
