% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sas.R
\name{check_az_installed}
\alias{check_az_installed}
\title{Internal function check_az_installed}
\usage{
check_az_installed()
}
\value{
Error message if az is not found or installed
}
\description{
Function that calls az on the command line to see if its installed
}
