% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rast_template.R
\name{create_rast_template}
\alias{create_rast_template}
\title{Load raster data}
\usage{
create_rast_template(aoi, crs, res, val = 1, save_dir = NULL)
}
\arguments{
\item{aoi}{sf geometry, area to be used to select pixels within the raster}

\item{crs}{Numeric, coordinate reference system as EPSG code}

\item{res}{Numeric, pixel resolution}

\item{val}{Numeric, value to fill the pixels}

\item{save_dir}{Character, path to save the raster}
}
\value{
SpatRast, with the provided specifications
}
\description{
Creates a raster grid from a geometry to be used for resampling other rasters
}
\examples{
\dontrun{
library(terra)
library(sf)

# Create AOI
coords <- base::matrix(c(
     -405086.7, 181125.8,
     -405086.7, 184647.8,
     -400843.1, 184647.8,
     -400843.1, 181125.8,
     -405086.7, 181125.8), 
     ncol = 2, byrow = TRUE)
aoi <- sf::st_polygon(list(coords)) |>
     sf::st_sfc(crs = 32198) |>
     sf::st_sf(crs = 32198)

# Create template
tmp_rast <- create_rast_template(aoi=aoi,crs=32198,res=50)

# Plot
plot(tmp_rast)
}
}
