% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sas.R
\name{create_sas}
\alias{create_sas}
\title{Internal function create_sas}
\usage{
create_sas(
  link_path = NULL,
  sas_duration = 180,
  retun_link_with_SAS = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{link_path}{Character, the http_link/path to file}

\item{sas_duration}{Double, number indicating the duration the SAS should allow access in minutes.}

\item{retun_link_with_SAS}{Boolean, whether to return the SAS token along, or appended with http link}

\item{verbose}{Boolean, to return full terminal output or hide it}
}
\value{
Character, sas token
}
\description{
Create a SAS token to be pasted inside your .Renvirob
}
\examples{
\dontrun{
# Create account sas
create_sas()
}
}
