% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_montreal_sf}
\alias{example_montreal_sf}
\title{Example Region over Montreal sf object}
\format{
sf object
}
\usage{
example_montreal_sf
}
\description{
A vector delimitation over montreal
}
\keyword{datasets}
