% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.move.R
\name{file.move}
\alias{file.move}
\title{Move files to another folder and rename them}
\source{
https://stackoverflow.com/questions/10266963/moving-files-between-folders
}
\usage{
file.move(from, to)
}
\arguments{
\item{from}{File path to be moved from the original location}

\item{to}{Destination path and file name}
}
\value{
Will only rename and/or move the file depending on the destination path and new file name if different from the file name of the 'from' argument.
}
\description{
Move files to another folder and rename them
}
\examples{
\dontrun{
# Generate a path to a file in temporary folder
f = file.path(tempdir(), c("original_file.txt"))
# Create the file
file.create(f)
# Move (rename) the file to another file name and/or location
file.move(from = f,
          to = file.path(tempdir(), c("new_file.txt")))
}
}
