% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Habitools-package.R
\docType{package}
\name{Habitools-package}
\alias{Habitools}
\alias{Habitools-package}
\title{Habitools: Tools for assisting in everyday analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Series of functions that makes it easy to load and prepare data from our internal cloud storage, with some additional utilities.
}
\keyword{internal}
