% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sas.R
\name{is.sas.valid}
\alias{is.sas.valid}
\title{Internal function is.sas.valid}
\usage{
is.sas.valid(sas_token)
}
\arguments{
\item{sas_token}{Character, sas token}
}
\value{
Character, sas token
}
\description{
Checks if the sas token in the .Renviron is valid
}
