% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raster.R
\name{load_raster}
\alias{load_raster}
\title{Load raster data}
\usage{
load_raster(COG_link = NULL, aoi = NULL, save_dir = NULL, desc = FALSE, ...)
}
\arguments{
\item{COG_link}{Character, a http link to a COG file}

\item{aoi}{sf geometry, area to be used to select pixels within the raster}

\item{save_dir}{Character, path to save the raster}

\item{desc}{If TRUE, will describe the raster (terra::describe)}

\item{...}{Additionnal arguments to terra::rast()}
}
\value{
SpatRast, containing pixels from the provided COG link within the AOI
}
\description{
Loads a COG file into memory using a http link and AOI for selecting pixels to load
}
\examples{
\dontrun{
library(terra)
library(sf)

# Create AOI
coords <- base::matrix(c(
     -405086.7, 181125.8,
     -405086.7, 184647.8,
     -400843.1, 184647.8,
     -400843.1, 181125.8,
     -405086.7, 181125.8),
     ncol = 2, byrow = TRUE)
aoi <- sf::st_polygon(list(coords)) |>
     sf::st_sfc(crs = 32198) |>
     sf::st_sf(crs = 32198)

# Grab Quebec LULC link
COG_link <- query_lookup(subject="utilisation-territoire",year=2020)$file_path

# Load LULC within the AOI
lulc <- load_raster(COG_link=COG_link,aoi=aoi)

# Plot
plot(lulc)
}
}
