% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_parquet.R
\name{open_parquet}
\alias{open_parquet}
\title{Open geoparquet in DuckDB}
\usage{
open_parquet(path_link, file_type = "parquet", link_expiration_time = 5)
}
\arguments{
\item{path_link}{Character, path/http_link to the parquet file to load}

\item{file_type}{Character, specify either "parquet" or "sf". All files from blob are parquet. sf is usefull for local files}

\item{link_expiration_time}{Double, duration of sas token for downloading (minutes)}
}
\value{
DuckDB
}
\description{
Streams geoparquet data from blob into local DuckDB instance for querying
}
\examples{
\dontrun{
# Grab parquet link
aoi_metadata <- query_lookup(country="canada",subject = "geo-boundaries-adm1-simplified")

# Open dataset
aoi<-open_parquet(aoi_metadata$file_path)

# View dataset
dplyr::glimpse(aoi)
aoi |>
    dplyr::distinct(shapeISO) |> 
    dplyr::pull()

# Filter DB and create sf object
filter_aoi<-aoi |>
    dplyr::filter(shapeISO == "CA-QB") |>
    to_sf(crs = 4326)

# Plot
plot(duckdbfs::to_sf(aoi,crs=4326)$geom)
plot(filter_aoi$geom,add=TRUE,col="red")
}
}
