% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_parquet.R
\name{prep_blob_paths}
\alias{prep_blob_paths}
\title{Internal function prep_blob_paths}
\usage{
prep_blob_paths(http_link, link_expiration_time)
}
\arguments{
\item{http_link}{Character, link to the parquet file to load}

\item{link_expiration_time}{Double, duration of sas token for downloading (minutes)}
}
\value{
Character, path of file
}
\description{
Prepares the file paths form the blob for DuckDB with SAS tokens added
}
