% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sas.R
\name{prepare_http_link}
\alias{prepare_http_link}
\title{Internal function prepare_http_link}
\usage{
prepare_http_link(link_path)
}
\arguments{
\item{link_path}{Character, the http_link/path to file}
}
\value{
List, name of the account, and path to file
}
\description{
Uses a http link to an asset on the blob to identify the storage account
}
