% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_fun.R
\name{process_raster}
\alias{process_raster}
\title{Change the format (CRS, extent, resolution, normalization) of an input raster to the desired properties}
\usage{
process_raster(
  input,
  roi_r,
  roi,
  project = TRUE,
  resample = TRUE,
  normalize = FALSE,
  scale = FALSE
)
}
\arguments{
\item{input}{Object with SpatRaster class or a Path to a raster (class character)}

\item{roi_r}{Raster of the ROI with CRS and used to resample}

\item{roi}{Polygon of the ROI}

\item{project}{Should reproject to the roi_r ?}

\item{resample}{Should resample to the roi_r ?}

\item{normalize}{Should normalize raster values ?}

\item{scale}{Should scale raster ?}
}
\value{
a Raster with the selected transformations
}
\description{
Generic function that takes input raster and prepares it for final data
package. Can reproject, resample, normalize (using the `scale_terra()`),
standardize (z-score with `terra::scale()` function).
}
\examples{
f <- system.file("ex/elev.tif", package="terra")
r <- terra::rast(f)
}
