% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_lookup.R
\name{query_lookup}
\alias{query_lookup}
\title{Query habitat's lookup table}
\usage{
query_lookup(
  complet_file_name = NULL,
  theme = NULL,
  country = NULL,
  region = NULL,
  subject = NULL,
  year = NULL,
  version = NULL,
  view_catalog = FALSE
)
}
\arguments{
\item{complet_file_name}{Character, string to filter the file name of the assets.}

\item{theme}{Character, string to filter the theme of the assets. (e.g. hydrology, biodiversity)}

\item{country}{Character, string to filter the country of the assets. (e.g. global, canada, america)}

\item{region}{Character, string to filter the region code of the assets. (e.g. all, qc)}

\item{subject}{Character, name of the file on the blob}

\item{year}{Numeric, number to filter the year of the assets. (e.g. 2023)}

\item{version}{Character, string to filter the version of the assets. (e.g. v0)}

\item{view_catalog}{Boolean, to return the names of the assets or not}
}
\value{
Dataframe, containing a single row with the filtered assests metadata
}
\description{
Uses the Habitat Blob storage lookup table and filters the rows to grab an asset link
}
\examples{
\dontrun{
library(dplyr)
geoboundaries_France <- query_lookup(country="france",subject = "geo-boundaries-adm1-simplified")
geoboundaries_France$file_path
}
}
