% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_parquet.R
\name{return_temp_path}
\alias{return_temp_path}
\title{Internal function return_temp_path}
\usage{
return_temp_path(x)
}
\arguments{
\item{x}{R object to save to disk}
}
\value{
Character, path to temp file
}
\description{
Takes a in memory object, saves it to disk, and returns the path to temp file
}
