% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_lookup.R
\name{rv_space_http}
\alias{rv_space_http}
\title{Internal function rv_space_http}
\usage{
rv_space_http(df)
}
\arguments{
\item{df}{Dataframe, Habitat lookup table containing http links}
}
\value{
Dataframe with white spaces removed in the http links
}
\description{
Removes white spaces in http links
}
