% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_fun.R
\name{scale_terra}
\alias{scale_terra}
\title{Rescale the raster values between 0 and 1.}
\source{
https://gis.stackexchange.com/questions/437520/normalize-raster-in-r
}
\usage{
scale_terra(r)
}
\arguments{
\item{r}{A raster}
}
\value{
raster values between 0 and 1
}
\description{
Uses the data in the raster (min and max) to transform the scale of the
raster to be between 0 and 1 following the equation (r - min) / (max - min).
}
\examples{
f <- system.file("ex/elev.tif", package="terra")
r <- terra::rast(f)
scale_terra(r)
}
