% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sas.R
\name{set_sas}
\alias{set_sas}
\title{Create SAS for .Renviron}
\usage{
set_sas(sas_duration = 180, update = FALSE, verbose = FALSE)
}
\arguments{
\item{sas_duration}{Double, number indicating the duration the SAS should allow access in minutes.}

\item{update}{Boolean, whether to update the sas token.}

\item{verbose}{Boolean, to return full terminal output or hide it.}
}
\value{
Character, messaging to update or check validity of sas in .Renviron
}
\description{
Create a SAS token to be pasted inside your .Renviron
}
\examples{
\dontrun{
# Create sas for .Renviron
set_sas(sas_duration=180)

# Update sas for .Renviron
set_sas(sas_duration=180,update=TRUE)
}
}
