% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_habitat.R
\name{theme_habitat}
\alias{theme_habitat}
\title{ggplot2 theme for habitat}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\value{
Themes for publication
}
\description{
Set of predefined themes for plotting
}
\details{
Loads a set of themes for ggplot2


\describe{

\item{`theme_habitat()`}{
A clean theme without the fuss}

}
}
\examples{
\dontrun{
library(ggplot2)

# Default theme
ggplot(iris) +
  geom_point(mapping = aes(x = Sepal.Length,
                           y = Sepal.Width,
                           color = Species))
# With theme_habitat presets
ggplot(iris) +
  geom_point(mapping = aes(x = Sepal.Length,
                           y = Sepal.Width,
                           color = Species)) +
  theme_habitat()
}
}
